Fixed compiler warnings. Explicit cast on crc vars (64 bit to 32 bit).  Need to be 64 bit for crcq instruction, but only 32 bits are used.

+++ Kernel/crc32c.c
@@ -194,8 +194,8 @@ uint32_t crc32c(uint32_t crc,const void * buf,size_t len)
                     : "r"(next), "0"(crc0), "1"(crc1), "2"(crc2));
             next += 8;
         } while (next < end);
-        crc0 = crc32c_shift(crc32c_long, crc0) ^ crc1;
-        crc0 = crc32c_shift(crc32c_long, crc0) ^ crc2;
+        crc0 = crc32c_shift(crc32c_long, (uint32_t)crc0) ^ crc1;
+        crc0 = crc32c_shift(crc32c_long, (uint32_t)crc0) ^ crc2;
         next += LONG*2;
         len -= LONG*3;
     }
@@ -214,8 +214,8 @@ uint32_t crc32c(uint32_t crc,const void * buf,size_t len)
                     : "r"(next), "0"(crc0), "1"(crc1), "2"(crc2));
             next += 8;
         } while (next < end);
-        crc0 = crc32c_shift(crc32c_short, crc0) ^ crc1;
-        crc0 = crc32c_shift(crc32c_short, crc0) ^ crc2;
+        crc0 = crc32c_shift(crc32c_short, (uint32_t)crc0) ^ crc1;
+        crc0 = crc32c_shift(crc32c_short, (uint32_t)crc0) ^ crc2;
         next += SHORT*2;
         len -= SHORT*3;
     }
